Fix brand icon color problem.
