+++ graylog2-web-interface/src/components/streamrules/StreamRuleModal.tsx
@@ -27,7 +27,7 @@ import {
   Select,
   ModalSubmit,
   InputOptionalInfo,
-  Spinner,
+  Spinner, BrandIcon,
 } from 'components/common';
 import HumanReadableStreamRule from 'components/streamrules//HumanReadableStreamRule';
 import { Col, Well, Input, Modal, Row } from 'components/bootstrap';
@@ -199,7 +199,7 @@ const StreamRuleModal = ({
                       The server will try to convert to strings or numbers based on the matcher type as well as it can.
 
                       <br /><br />
-                      <Icon name="code" type="brand" />&nbsp;
+                      <BrandIcon name="github" />&nbsp;
                       <a href={`https://github.com/Graylog2/graylog2-server/tree/${Version.getMajorAndMinorVersion()}/graylog2-server/src/main/java/org/graylog2/streams/matchers`}
                          target="_blank"
                          rel="noopener noreferrer"> Take a look at the matcher code on GitHub
